hp줄어들고 아이템 먹으면 hp 참

+++ (Gamp)Project/Project/main.cpp
@@ -9,7 +9,7 @@ int main()
 	PLAYER player = {};
 	POS startpos = {};
 	Init(map, &player, &startpos);
-	int hpCnt = 31;
+	float hpCnt = 31;
 
 	while (true)
 	{
